Rename NetSuite REST adaptor to just NetSuite
