+++ lib/ledger_sync/adaptors/netsuite/customer/operations/update.rb
@@ -1,0 +1,24 @@
+# frozen_string_literal: true
+
+module LedgerSync
+  module Adaptors
+    module NetSuite
+      module Customer
+        module Operations
+          class Update < NetSuite::Operation::Update
+            class Contract < LedgerSync::Adaptors::Contract
+              params do
+                required(:external_id).maybe(:string)
+                required(:ledger_id).filled(:string)
+                required(:email).maybe(:string)
+                required(:name).filled(:string)
+                required(:phone_number).maybe(:string)
+                required(:subsidiary).filled(:hash, Types::Reference)
+              end
+            end
+          end
+        end
+      end
+    end
+  end
+end
