--- lib/ledger_sync/adaptors/netsuite_rest/customer/operations/delete.rb
@@ -1,24 +1,0 @@
-# frozen_string_literal: true
-
-module LedgerSync
-  module Adaptors
-    module NetSuiteREST
-      module Customer
-        module Operations
-          class Delete < NetSuiteREST::Operation::Delete
-            class Contract < LedgerSync::Adaptors::Contract
-              params do
-                required(:external_id).maybe(:string)
-                required(:ledger_id).filled(:string)
-                required(:email).maybe(:string)
-                required(:name).maybe(:string)
-                required(:phone_number).maybe(:string)
-                required(:subsidiary).maybe(:hash, Types::Reference)
-              end
-            end
-          end
-        end
-      end
-    end
-  end
-end
