+++ spec/integrations/netsuite_soap/customers/create_spec.rb
@@ -35,7 +35,7 @@ RSpec.describe LedgerSync::Adaptors::NetSuiteSOAP::Customer::Operations::Create,
 
     it { expect_valid(operation: initialized_operation) }
     it 'creates', vcr: true do
-      VCR.use_cassette('netsuite/customers/create') do
+      VCR.use_cassette('netsuite_soap/customers/create') do
         expect(subject).to be_a(LedgerSync::OperationResult::Success)
         expect(subject.resource.ledger_id).to eq('123')
       end
